backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
stripe>=7.0.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
slowapi>=0.1.9
bleach>=6.1.0
resend>=0.8.0
mailjet-rest>=1.3.4
supabase>=2.3.0
PyJWT[crypto]>=2.8.0
requests>=2.31.0
psycopg2-binary>=2.9.9
